+++ test/index.js
@@ -2,38 +2,10 @@ import cpreact, { Component, ReactDOM } from '../src/index'
 
 // You can try to run test demo in this file
 // how to use ? you can read https://github.com/MuYunyun/cpreact/blob/master/.github/PULL_REQUEST_TEMPLATE.md
-// class App extends Component {
-//   render() {
-//     return (
-//       <div>hello, cpreact</div>
-//     )
-//   }
-// }
 class App extends Component {
-  constructor(props) {
-    super(props)
-    this.state = {
-      count: 0
-    }
-    this.click = this.click.bind(this)
-  }
-
-  click() {
-    for (let i = 0; i < 10; i++) {
-      this.setState({ // 在先前的逻辑中，没调用一次 setState 就会 render 一次
-        count: ++this.state.count
-      })
-    }
-
-    this.state.count = 13;
-  }
-
   render() {
     return (
-      <div>
-        <button onClick={this.click}>增加</button>
-        <div>{this.state.count}</div>
-      </div>
+      <div>hello, cpreact</div>
     )
   }
 }